Update DivisionGame.c
A program is required for a computer game. The user keys in the number of rounds he wishes toplay.For each round the user enters his lucky number. The program takes the number and divides it withasecret number. If the remainder of the division is zero, it is considered a draw for the roundandthetotalscore is incriminated by 1. Otherwise if it is any other even number, it is considered a winfor theroundand the total score is incremented by 3. However if it is an odd number, it is considered a loss fortheround and the total score is decremented by 3. This is done until he completes his rounds. Hewinsifthetotal score at the end is a positive number.