Create GradingSystem.c
Write a program that accepts the student admission number, student surname and three subjectsmarksanddisplays the same along with the student average mark. The program should also display the student grade depending on the average marks the student obtains.The grading system is shown in the table below
Average Mark Grade
0-39 Fail
40-50 D
50-59 C
60-69 B
70-100 A
The program should also prompt the user if to continue. If the user enters “Y”, for yes thenprogramcomputes for another student and “N”, for no then program exits