Create CreditProgram.c
A computer program is required to evaluate the credit worthy of a client. The programreads thecreditlimit, price and quantity of the item to be purchased by the client. If the value of the goods is morethanthe credit limit, the program displays “Sorry you cannot purchase goods worthy such a valueoncredit”and allows the customer to re-enter the quantity, otherwise, displays “Thank You for purchasingfromus”and the value of the purchase. This should be repeated for n customers